requirements.txt
Django==2.0.6
PyJWT==1.6.4
djangorestframework==3.8.2
